Moved RPC related types to a dedicated package
